fix: resolve composer tests collection error (file/directory conflict)
Move tests.py content into tests/__init__.py and delete the conflicting
file. Python can't resolve the module when both tests.py and tests/
exist with the same basename.
